% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeExplorer.R
\name{cnvVis}
\alias{cnvVis}
\title{Creates a configuration for "Copy number variation" visualization}
\usage{
cnvVis(seg = NULL, baf = NULL)
}
\arguments{
\item{seg}{Name of the segment mean variable}

\item{baf}{Name of the B allele frequency variable}
}
\value{
A configuration for the cnv visualization
}
\description{
Both of the parameters \code{seg} and \code{baf} are optional, but leaving
both empty is pointless.
}
