% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeExplorer.R
\name{readCytobands}
\alias{readCytobands}
\title{Reads a cytobands file into a data frame}
\usage{
readCytobands(file)
}
\arguments{
\item{file}{Path to the cytoband file}
}
\description{
This function can be used to load a cytoband file for a genome assembly
that is not supported out of the box by the GenomeExplorer. A loaded
data frame can be fed to GenomeExplorer instead of a name of a genome assembly.
}
\details{
An example of a cytoband file:
http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/cytoBand.txt.gz

Cytoband files for some genome assemblies are included in the extdata directory
}
