% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeSpy.R
\name{refseqGeneTrack}
\alias{refseqGeneTrack}
\title{Adds a refseq gene annotation track to Genome Spy}
\usage{
refseqGeneTrack(ge, refseq_genes_compressed = NULL)
}
\arguments{
\item{refseq_genes_compressed}{A URL to "delta compressed" refseq genes. See utils/compressRefGene.py

  The server must allow cross origin requests by including the following HTTP header in responses:
  \code{Access-Control-Allow-Origin: *}
  
  If the URL is omitted, the genome data will be loaded from karilavikka.fi.}
}
\description{
Adds a refseq gene annotation track to Genome Spy
}
