% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeSpy.R
\name{search}
\alias{search}
\title{Searches a cytoband, chromosome or a refseq gene.}
\usage{
search(ge, search)
}
\arguments{
\item{ge}{A GenomeSpy instance}

\item{search}{The search keyword}
}
\description{
Search accepts the same search keywords and formats as the search field in
the user interface.
}
\details{
Gene search is only available if gene annotation track has been added to the
Spy.
}
\examples{
search(ge, "tp53")
search(ge, "14q32")

}
