% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeExplorer.R
\name{segmentTrack}
\alias{segmentTrack}
\title{Adds a segment track to Genome Explorer}
\usage{
segmentTrack(ge, data, discriminator = NULL, chrom = "chrom",
  start = "start", end = "end", samples = NULL, sample_labels = NULL,
  vis = "simple", vis_config = list(), title = NULL, subtrack_size = 30,
  subtrack_padding = 0.2)
}
\arguments{
\item{ge}{The Genome Explorer object.}

\item{data}{A data frame that contains the segments being visualized.}

\item{discriminator}{An optional discriminator column that is used for
spreading the data (cases) to subtracks.}

\item{chrom}{The column containing the name of the chromosome.}

\item{start}{The column containing the start coordinate of the genomic region.}

\item{end}{The column containing the end coordinate of the genomic region.}

\item{samples}{An optional vector that specifies the order of the samples.
By default the samples are ordered alphabetically.}

\item{sample_labels}{An optional vector that specifies replacement labels
for the samples.}

\item{vis}{Visualization to use.}

\item{vis_config}{Optional configuration options for the visualization.}

\item{title}{Title of the track.}

\item{subtrack_size}{Subtrack thickness (height) in pixels.}

\item{subtrack_padding}{Padding between the subtracks in proportion to thickness.}
}
\description{
A segment is a region in a chromosome. It has start and end coordinates, which are
indexed in a UCSC style 1-indexed closed ranges.
}
\details{
A segment track uses a user-defined method for visualizing arbitrary columns
in the data. Currently supported methods are \code{simple} and \code{cnvAndBaf}.
}
