% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeExplorer.R
\name{zoom}
\alias{zoom}
\title{Zooms into a range}
\usage{
zoom(ge, search)
}
\arguments{
\item{search}{Search keyword to use for zooming in}
}
\description{
Currently the range must be specified as a search string, which accepts
the same search keywords and formats as the search field in the user interface.
}
