% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeExplorer.R
\name{zoom}
\alias{zoom}
\title{Zooms into the specified region}
\usage{
zoom(ge, startChrom, startPos, endChrom, endPos)
}
\arguments{
\item{ge}{A GenomeExplorer instance}

\item{startChrom, endChrom}{A chromosome}

\item{startPos, endPos}{A position within the chromosome}
}
\description{
Zooms into the specified region
}
\examples{
zoom(ge, "chr2", 123456, "chr4", 34567890)

}
